Modify edit command classes code structure
