+++ src/main/java/seedu/duke/commands/EditAudioCommand.java
@@ -112,18 +112,16 @@ public class EditAudioCommand extends Command {
     }
 
     /**
-     * Executes Edit Audio command.
-     * Overrides method from parent class.
+     * Processes Edit Audio Command, including exceptions.
      *
      * @param ui Object that handles user IO
      * @param catalogue Object that encapsulates the library catalogue
+     * @throws LibmgrException when the user input is invalid
      */
-    @Override
-    public void execute(TextUI ui, Catalogue catalogue) {
+    public void handlesEditAudioCommand(TextUI ui, Catalogue catalogue) throws LibmgrException {
         processArgs();
         if (checkMissingArgs() || checkEmptyArgs()) {
-            ui.print(EDIT_AUDIO_INVALID_FORMAT);
-            return;
+            throw new LibmgrException(EDIT_AUDIO_INVALID_FORMAT);
         }
         if (checkInvalidArgs()) {
             ui.print(WARN_INVALID_ARGS);
@@ -131,42 +129,49 @@ public class EditAudioCommand extends Command {
         if (args.containsKey(KEY_TITLE)) {
             assert title != null && !title.equals("");
             if (toEdit.getTitle().equals(title)) {
-                ui.print(EDIT_UNCHANGED_TITLE);
-                return;
+                throw new LibmgrException(EDIT_UNCHANGED_TITLE);
             }
             toEdit.setTitle(title);
         }
         if (args.containsKey(KEY_ID)) {
             assert id != null && !id.equals("");
             if (toEdit.getID().equals(id)) {
-                ui.print(EDIT_UNCHANGED_ID);
-                return;
+                throw new LibmgrException(EDIT_UNCHANGED_ID);
             }
-            try {
-                catalogue.checkDuplicateID(id);
-            } catch (LibmgrException e) {
-                ui.print(e.getMessage());
-                return;
-            }
+            catalogue.checkDuplicateID(id);
             toEdit.setID(id);
         }
         if (args.containsKey(KEY_ARTIST)) {
             assert artist != null && !artist.equals("");
             if (toEdit.getArtist().equals(artist)) {
-                ui.print(EDIT_UNCHANGED_ARTIST);
-                return;
+                throw new LibmgrException(EDIT_UNCHANGED_ARTIST);
             }
             toEdit.setArtist(artist);
         }
         if (args.containsKey(KEY_DURATION)) {
             assert duration != null && !duration.equals("");
             if (toEdit.getDuration().equals(duration)) {
-                ui.print(EDIT_UNCHANGED_DURATION);
-                return;
+                throw new LibmgrException(EDIT_UNCHANGED_DURATION);
             }
             toEdit.setDuration(duration);
         }
         ui.print(EDIT_AUDIO_MESSAGE, toEdit);
+    }
+
+    /**
+     * Executes Edit Audio command, including exception handling
+     * Overrides method from parent class.
+     *
+     * @param ui Object that handles user IO
+     * @param catalogue Object that encapsulates the library catalogue
+     */
+    @Override
+    public void execute(TextUI ui, Catalogue catalogue) {
+        try {
+            handlesEditAudioCommand(ui, catalogue);
+        } catch (LibmgrException e) {
+            ui.print(e.getMessage());
+        }
     }
 }
 //@@author avellinwong01
