+++ src/main/java/seedu/duke/commands/EditMagazineCommand.java
@@ -112,18 +112,16 @@ public class EditMagazineCommand extends Command {
     }
 
     /**
-     * Executes Edit Magazine command.
-     * Overrides method from parent class.
+     * Processes Edit Magazine Command, including exceptions.
      *
      * @param ui Object that handles user IO
      * @param catalogue Object that encapsulates the library catalogue
+     * @throws LibmgrException when the user input is invalid
      */
-    @Override
-    public void execute(TextUI ui, Catalogue catalogue) {
+    public void handlesEditMagazineCommand(TextUI ui, Catalogue catalogue) throws LibmgrException {
         processArgs();
         if (checkMissingArgs() || checkEmptyArgs()) {
-            ui.print(EDIT_MAGAZINE_INVALID_FORMAT);
-            return;
+            throw new LibmgrException(EDIT_MAGAZINE_INVALID_FORMAT);
         }
         if (checkInvalidArgs()) {
             ui.print(WARN_INVALID_ARGS);
@@ -131,42 +129,49 @@ public class EditMagazineCommand extends Command {
         if (args.containsKey(KEY_TITLE)) {
             assert title != null && !title.equals("");
             if (toEdit.getTitle().equals(title)) {
-                ui.print(EDIT_UNCHANGED_TITLE);
-                return;
+                throw new LibmgrException(EDIT_UNCHANGED_TITLE);
             }
             toEdit.setTitle(title);
         }
         if (args.containsKey(KEY_ID)) {
             assert id != null && !id.equals("");
             if (toEdit.getID().equals(id)) {
-                ui.print(EDIT_UNCHANGED_ID);
-                return;
+                throw new LibmgrException(EDIT_UNCHANGED_ID);
             }
-            try {
-                catalogue.checkDuplicateID(id);
-            } catch (LibmgrException e) {
-                ui.print(e.getMessage());
-                return;
-            }
+            catalogue.checkDuplicateID(id);
             toEdit.setID(id);
         }
         if (args.containsKey(KEY_PUBLISHER)) {
             assert publisher != null && !publisher.equals("");
             if (toEdit.getPublisher().equals(publisher)) {
-                ui.print(EDIT_UNCHANGED_PUBLISHER);
-                return;
+                throw new LibmgrException(EDIT_UNCHANGED_PUBLISHER);
             }
             toEdit.setPublisher(publisher);
         }
         if (args.containsKey(KEY_EDITION)) {
             assert edition != null && !edition.equals("");
             if (toEdit.getEdition().equals(edition)) {
-                ui.print(EDIT_UNCHANGED_EDITION);
-                return;
+                throw new LibmgrException(EDIT_UNCHANGED_EDITION);
             }
             toEdit.setEdition(edition);
         }
         ui.print(EDIT_MAGAZINE_MESSAGE, toEdit);
+    }
+
+    /**
+     * Executes Edit Magazine command, including exception handling.
+     * Overrides method from parent class.
+     *
+     * @param ui Object that handles user IO
+     * @param catalogue Object that encapsulates the library catalogue
+     */
+    @Override
+    public void execute(TextUI ui, Catalogue catalogue) {
+        try {
+            handlesEditMagazineCommand(ui, catalogue);
+        } catch (LibmgrException e) {
+            ui.print(e.getMessage());
+        }
     }
 }
 //@@author avellinwong01
